image service filename

+++ src/main/java/com/webapp/demo/service/imageService.java
@@ -17,7 +17,7 @@ public class ImageService {
 
     //@Override
     @Transactional
-    public Products saveImageFile(Long id, MultipartFile file) {
+    public Products saveImageFile(int id, MultipartFile file) {
 
         try {
             Products product = productsRepo.findById(Math.toIntExact(id)).get();
